Upgrade and cleanup
