+++ assets/demo.rml
@@ -7,15 +7,15 @@
 			{
 				width: 300dp;
 				height: 225dp;
-				
+
 				margin: auto;
 			}
-			
+
 			div#title_bar div#icon
 			{
 				display: none;
 			}
-			
+
 			div#content
 			{
 				text-align: left;
