+++ assets/rml.rcss
@@ -27,12 +27,7 @@ strong
 	font-weight: bold;
 }
 
-datagrid
-{
-	display: block;
-}
-
-select, dataselect, datacombo
+select
 {
 	text-align: left;
 }
