+++ src/samples/basic/bitmapfont/data/bitmapfont.rml
@@ -1,0 +1,31 @@
+<rml>
+<head>
+<title>Demo</title>
+<link type="text/template" href="../../../assets/window.rml" />
+<style>
+body
+{
+	width: 600px;
+	height: 500px;
+	font-family: Comfortaa;
+	margin: auto;
+	font-size: 22px;
+}
+
+div#title_bar div#icon
+{
+	display: none;
+}
+
+div#content
+{
+	text-align: left;
+}
+</style>
+</head>
+<body template="window">
+	<p>This is a bitmap font sample!</p>
+	<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Urna neque viverra justo nec ultrices dui sapien eget mi. Risus quis varius quam quisque id. Amet est placerat in egestas erat imperdiet. Velit egestas dui id ornare arcu odio ut sem. Aliquet porttitor lacus luctus accumsan tortor posuere. Et malesuada fames ac turpis egestas integer eget. Enim nunc faucibus a pellentesque sit amet porttitor eget. Nunc pulvinar sapien et ligula. Sit amet mattis vulputate enim nulla aliquet porttitor lacus luctus. Dolor sit amet consectetur adipiscing. Congue eu consequat ac felis donec et odio pellentesque. Nunc non blandit massa enim nec dui nunc mattis.</p>
+	<p>Posuere sollicitudin aliquam ultrices sagittis orci a scelerisque. Quam pellentesque nec nam aliquam sem et. Et tortor consequat id porta nibh venenatis. Facilisi cras fermentum odio eu feugiat pretium nibh ipsum consequat. Magna ac placerat vestibulum lectus mauris ultrices eros. Amet purus gravida quis blandit turpis cursus. Elit duis tristique sollicitudin nibh sit amet commodo nulla. Suscipit adipiscing bibendum est ultricies integer quis auctor elit sed. Iaculis at erat pellentesque adipiscing commodo elit at imperdiet. Pellentesque elit ullamcorper dignissim cras. Nunc eget lorem dolor sed viverra ipsum nunc. Ac felis donec et odio pellentesque diam volutpat commodo.</p>
+</body>
+</rml>
